Overhaul
Instead of relying on modifying the code to point to the folders and
mappings you need, I made it prompt you for it instead.

This also means I added functionality that tests/checks your mapping and
CFC folder to see if they exist and also correspond with each other.

There's a few other fixes I forgot about.  Also removed a lot of the
jQuery files that were not in use.